requirements.txt
torch
torchvision
tqdm
scipy
tensorboard
mlflow
faiss-cpu